OCR Methods
Added OCR methods and updated requirements list to reflect proper methods.

Unfortunately, the current workflow is rather slow, but it uses pytesseract to convert a given PDF to one embedded with proper OCR, a text file, or both depending on the method selected.

I will try to play with it some more to get it to be faster, but I wanted to make sure I got something up before the break.