Added deskew feature and added ocr folder
The OCR methods will take up a bit of space with some of the different files, so I've added a seperate folder under djweb.

Additionally, I think I have found a function that may be able to deskew our files. Will update later.

I'm commiting this so Meesue is able to look at the file.