Generates Dictionaries of Facets of an Object in Advanced Search
Method creates a dictionary called dict_facets. The keys of dict_facets are the different kinds of facets.  These keys’ values are dictionaries. Within these facet dictionaries, the keys are the top 10 most common instances of the facet and the values are integers that signify how often that instance occurs in the metadata.

Example of Organization:

dict_facets={
            "authors":  {"Philip M. Morse”:1}
	“author organizations”: {“MIT”:1}
	“recipients”: {“N. M. Sage”:1}
	“recipient organizations”: {“Committee on Machine Computation”:1}
	“cceds”: {“J. W. Forrester”:1}
	“cced organizations”: {“LWH”:1}
           "years": {1950:1}
}