Display of Object's Facets in Advanced Search
When a person searches something in advanced search, this will display the top 10 most occurent instance of every facet (authors, author organizations, recipients, recipient organizations, cceds, cced organizations, year) of that object.

We can use this as reference when analyzing. It would also be interesting for the user to see the information.