Updated metadata
Some of the document ids were generated based on the doc_ids in the row above. We recently sorted metadata by author, which meant that documents didn't appear by folder anymore, meaning that document ids for some folders were essentially random.
All doc ids in the google sheet are now static and the csv in this PR populates without issues and should solve the problem.